requirements.txt
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.2
Pillow==10.3.0
pyttsx3==2.90
customtkinter==5.2.2